ExampleTask
This is an example of a task where nothing is executed, but I'll briefly explain the methods. 

ShouldRun is where you set the conditions for whether the task will run or not. If all conditions are met, "true" is returned; if none are met, "false" is returned.

Run is the logic to look for when the task runs; for example, if the villager dances or something similar.

ShouldKeepRunning indicates whether or not the task will complete in the next few ticks; for example, at a certain time when the villager finishes dancing. "False" is returned to terminate the task.